+++ nequip/ase/nequip_calculator.py
@@ -24,7 +24,7 @@ class NequIPCalculator(Calculator):
 
     """
 
-    implemented_properties = ["energy", "forces"]
+    implemented_properties = ["energy", "energies", "forces"]
 
     def __init__(
         self,
@@ -103,14 +103,19 @@ class NequIPCalculator(Calculator):
 
         # prepare data
         data = AtomicData.from_ase(atoms=atoms, r_max=self.r_max)
+        for k in AtomicDataDict.ALL_ENERGY_KEYS:
+            if k in data:
+                del data[k]
         data = self.transform(data)
-
         data = data.to(self.device)
+        data = AtomicData.to_AtomicDataDict(data)
 
         # predict + extract data
-        out = self.model(AtomicData.to_AtomicDataDict(data))
+        out = self.model(data)
         forces = out[AtomicDataDict.FORCE_KEY].detach().cpu().numpy()
-        energy = out[AtomicDataDict.TOTAL_ENERGY_KEY].detach().cpu().item()
+        energy = (
+            out[AtomicDataDict.TOTAL_ENERGY_KEY].detach().cpu().numpy().reshape(tuple())
+        )
 
         # store results
         self.results = {
@@ -118,3 +123,12 @@ class NequIPCalculator(Calculator):
             # force has units eng / len:
             "forces": forces * (self.energy_units_to_eV / self.length_units_to_A),
         }
+
+        if AtomicDataDict.PER_ATOM_ENERGY_KEY in out:
+            self.results["energies"] = self.energy_units_to_eV * (
+                out[AtomicDataDict.PER_ATOM_ENERGY_KEY]
+                .detach()
+                .squeeze(-1)
+                .cpu()
+                .numpy()
+            )