+++ tests/integration/test_deploy.py
@@ -9,7 +9,8 @@ import numpy as np
 import torch
 
 import nequip
-from nequip.data import AtomicDataDict, AtomicData
+from nequip.data import AtomicDataDict, AtomicData, dataset_from_config
+from nequip.utils import Config
 from nequip.scripts import deploy
 from nequip.train import Trainer
 from nequip.ase import NequIPCalculator
@@ -18,19 +19,26 @@ from nequip.ase import NequIPCalculator
 @pytest.mark.parametrize(
     "device", ["cpu"] + (["cuda"] if torch.cuda.is_available() else [])
 )
-def test_deploy(nequip_dataset, BENCHMARK_ROOT, device):
+def test_deploy(BENCHMARK_ROOT, device):
     dtype = str(torch.get_default_dtype())[len("torch.") :]
+    atol = {"float32": 1e-5, "float64": 1e-7}[dtype]
 
     # if torch.cuda.is_available():
     #     # TODO: is this true?
     #     pytest.skip("CUDA and subprocesses have issues")
+
+    keys = [
+        AtomicDataDict.TOTAL_ENERGY_KEY,
+        AtomicDataDict.FORCE_KEY,
+        AtomicDataDict.PER_ATOM_ENERGY_KEY,
+    ]
 
     config_path = pathlib.Path(__file__).parents[2] / "configs/minimal.yaml"
     true_config = yaml.load(config_path.read_text(), Loader=yaml.Loader)
     with tempfile.TemporaryDirectory() as tmpdir:
         # Save time
         run_name = "test_deploy" + dtype
-        root = "./"
+        root = tmpdir + "/nequip_rootdir/"
         true_config["run_name"] = run_name
         true_config["root"] = root
         true_config["dataset_file_name"] = str(
@@ -41,7 +49,8 @@ def test_deploy(nequip_dataset, BENCHMARK_ROOT, device):
         true_config["n_train"] = 1
         true_config["n_val"] = 1
         config_path = "conf.yaml"
-        with open(f"{tmpdir}/{config_path}", "w+") as fp:
+        full_config_path = f"{tmpdir}/{config_path}"
+        with open(full_config_path, "w+") as fp:
             yaml.dump(true_config, fp)
         # Train model
         retcode = subprocess.run(["nequip-train", str(config_path)], cwd=tmpdir)
@@ -58,18 +67,19 @@ def test_deploy(nequip_dataset, BENCHMARK_ROOT, device):
 
         # now test predictions the same
         best_mod, _ = Trainer.load_model_from_training_session(
-            traindir=f"{tmpdir}/{root}/{run_name}/",
+            traindir=f"{root}/{run_name}/",
             model_name="best_model.pth",
             device=device,
         )
         best_mod.eval()
 
-        data = AtomicData.to_AtomicDataDict(nequip_dataset[0].to(device))
-        # Needed because of debug mode:
-        data[AtomicDataDict.TOTAL_ENERGY_KEY] = data[
-            AtomicDataDict.TOTAL_ENERGY_KEY
-        ].unsqueeze(0)
-        train_pred = best_mod(data)[AtomicDataDict.TOTAL_ENERGY_KEY].to("cpu")
+        # load train dataset, already cached
+        dataset = dataset_from_config(Config.from_file(full_config_path))
+        data = AtomicData.to_AtomicDataDict(dataset[0].to(device))
+        for k in keys:
+            data.pop(k, None)
+        train_pred = best_mod(data)
+        train_pred = {k: train_pred[k].to("cpu") for k in keys}
 
         # load model and check that metadata saved
         # TODO: use both CPU and CUDA to load?
@@ -84,9 +94,13 @@ def test_deploy(nequip_dataset, BENCHMARK_ROOT, device):
         assert len(metadata[deploy.TYPE_NAMES_KEY].split(" ")) == 3  # C, H, O
 
         data_idx = 0
-        data = AtomicData.to_AtomicDataDict(nequip_dataset[data_idx].to("cpu"))
-        deploy_pred = deploy_mod(data)[AtomicDataDict.TOTAL_ENERGY_KEY]
-        assert torch.allclose(train_pred, deploy_pred, atol=1e-7)
+        data = AtomicData.to_AtomicDataDict(dataset[data_idx].to("cpu"))
+        for k in keys:
+            data.pop(k, None)
+        deploy_pred = deploy_mod(data)
+        deploy_pred = {k: deploy_pred[k].to("cpu") for k in keys}
+        for k in keys:
+            assert torch.allclose(train_pred[k], deploy_pred[k], atol=atol)
 
         # now test info
         # hack for old version
@@ -109,9 +123,22 @@ def test_deploy(nequip_dataset, BENCHMARK_ROOT, device):
             deployed_path,
             device="cpu",
             species_to_type_name={s: s for s in ("C", "H", "O")},
+            set_global_options=False,
         )
         # use .get() so it's not transformed
-        atoms = nequip_dataset.get(data_idx).to_ase()
+        atoms = dataset.get(dataset.indices()[data_idx]).to_ase()
         atoms.calc = calc
-        ase_forces = atoms.get_potential_energy()
-        assert torch.allclose(train_pred, torch.as_tensor(ase_forces), atol=1e-7)
+        ase_pred = {
+            AtomicDataDict.TOTAL_ENERGY_KEY: atoms.get_potential_energy(),
+            AtomicDataDict.FORCE_KEY: atoms.get_forces(),
+            AtomicDataDict.PER_ATOM_ENERGY_KEY: atoms.get_potential_energies(),
+        }
+        assert ase_pred[AtomicDataDict.TOTAL_ENERGY_KEY].shape == tuple()
+        assert ase_pred[AtomicDataDict.FORCE_KEY].shape == (len(atoms), 3)
+        assert ase_pred[AtomicDataDict.PER_ATOM_ENERGY_KEY].shape == (len(atoms),)
+        for k in keys:
+            assert torch.allclose(
+                deploy_pred[k].squeeze(-1),
+                torch.as_tensor(ase_pred[k], dtype=torch.get_default_dtype()),
+                atol=atol,
+            )